moved qrgenerator to helpers
