+++ RefrigerantCylinderScanner/Helpers/QrGenerator.swift
@@ -1,0 +1,27 @@
+//
+//  QrGenerator.swift
+//  RefrigerantCylinderScanner
+//
+//  Created by Dave Szczutkowski on 03/08/2022.
+//
+
+import Foundation
+import CoreImage
+import CoreImage.CIFilterBuiltins
+import SwiftUI
+
+struct QrGenerator {
+    let context = CIContext()
+    let filter = CIFilter.qrCodeGenerator()
+    
+    func generateQrCode(from string: String) -> UIImage {
+        filter.message = Data(string.utf8)
+        
+        if let outputImage = filter.outputImage {
+            if let cgImage = context.createCGImage(outputImage, from: outputImage.extent) {
+                return UIImage(cgImage: cgImage)
+            }
+        }
+        return UIImage(systemName: "xmark.circle") ?? UIImage()
+    }
+}
